refactor: Remove remaining departures v1 stuff (#4115)
