--- src/quay-departures-screen/index.tsx
@@ -1,4 +1,0 @@
-export {QuayDeparturesScreenComponent} from './QuayDeparturesScreenComponent';
-export type {QuayDeparturesScreenParams} from './QuayDeparturesScreenComponent';
-export {useDepartureData} from './use-departure-data';
-export type {SearchTime} from './types';
