--- src/quay-departures-screen/types.ts
@@ -1,6 +1,0 @@
-export const DateOptions = ['now', 'departure'] as const;
-export type DateOptionType = (typeof DateOptions)[number];
-export type SearchTime = {
-  option: DateOptionType;
-  date: string;
-};
