+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_DashboardStack/navigation-types.ts
@@ -2,7 +2,6 @@ import {DepartureDetailsScreenParams} from '@atb/travel-details-screens/Departur
 import {TravelDetailsMapScreenParams} from '@atb/travel-details-map-screen';
 import {PlaceScreenParams} from '@atb/place-screen/PlaceScreenComponent';
 import {TripDetailsScreenParams} from '@atb/travel-details-screens/TripDetailsScreenComponent';
-import {QuayDeparturesScreenParams} from '@atb/quay-departures-screen';
 import {JourneyDatePickerScreenParams} from '@atb/journey-date-picker';
 import {TabNavigatorScreenProps} from '@atb/stacks-hierarchy/Root_TabNavigatorStack';
 import {
@@ -31,7 +30,6 @@ export type DashboardStackParams = {
   Dashboard_PlaceScreen: Dashboard_PlaceScreenParams;
   Dashboard_TripSearchScreen: TripSearchScreenParams;
   Dashboard_TripDetailsScreen: TripDetailsScreenParams;
-  Dashboard_QuayDeparturesScreen: QuayDeparturesScreenParams;
   Dashboard_JourneyDatePickerScreen: JourneyDatePickerScreenParams;
   Dashboard_FavoriteDeparturesScreen: undefined;
   Dashboard_NearbyStopPlacesScreen: Dashboard_NearbyStopPlacesScreenParams;
