+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_MapStack/navigation-types.ts
@@ -1,7 +1,6 @@
 import {PlaceScreenParams} from '@atb/place-screen/PlaceScreenComponent';
 import {DepartureDetailsScreenParams} from '@atb/travel-details-screens/DepartureDetailsScreenComponent';
 import {TravelDetailsMapScreenParams} from '@atb/travel-details-map-screen';
-import {QuayDeparturesScreenParams} from '@atb/quay-departures-screen';
 import {TabNavigatorScreenProps} from '@atb/stacks-hierarchy/Root_TabNavigatorStack';
 import {CompositeScreenProps} from '@react-navigation/native';
 import {StackScreenProps} from '@react-navigation/stack';
@@ -12,7 +11,6 @@ export type MapStackParams = {
   Map_PlaceScreen: PlaceScreenParams;
   Map_DepartureDetailsScreen: DepartureDetailsScreenParams;
   Map_TravelDetailsMapScreen: TravelDetailsMapScreenParams;
-  Map_QuayDeparturesScreen: QuayDeparturesScreenParams;
 };
 
 type RootMapScreenProps = TabNavigatorScreenProps<'TabNav_MapStack'>;
